remove policy
