--- tests/Kros.Ocelot.ETagCaching.Test/ETagCacheContextShould.OverrideToString.received.txt
@@ -1,11 +1,0 @@
-﻿ETagCacheContext
-{
-	ETag: "default"
-	CacheKey: ""
-	EnableETagCache: False
-	AllowNotModified: False
-
-	ETagExpirationTimeSpan: "00:00:00"
-	Tags: []
-	StatusCode: 0
-}
